+++ src/main/java/rench/backend/models/Country.java
@@ -13,14 +13,14 @@ import java.util.List;
 public class Country {
 
     public Country() { }
-    public Country(Long id) {
+    public Country(int id) {
         this.id = id;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", updatable = false, nullable = false)
-    public long id;
+    public int id;
 
     @Column(name = "name", nullable = false, unique = true)
     public String name;
