+++ src/main/java/rench/backend/repositories/CountryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import rench.backend.models.Country;
 
 @Repository
-public interface CountryRepository  extends JpaRepository<Country, Long>
+public interface CountryRepository  extends JpaRepository<Country, Integer>
 {
 
 }
