+++ src/main/java/rench/backend/repositories/MuseumRepository.java
@@ -3,6 +3,5 @@ package rench.backend.repositories;
 import rench.backend.models.Museum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MuseumRepository extends JpaRepository<Museum, Long> {
+public interface MuseumRepository extends JpaRepository<Museum, Integer> {
 }
-
